Move playback and midiClips into separate files
